Add a hardware interface, I_PLLSynthesizer
The methods used by the MAX2871 and ADF4356 are very close to each
other but not exact. By providing an interface it will make it
much easier to swap between synthesizer chips or to use multiple
different chips in the same design. The interface, I_PLLSynthesizer
is a pure virtual class which means you only need to implement the
same functions in your driver code.