Refactor outputSelect enable/disable set/clear
Originally the function used negative logic, e.g. disableA and
disableB with the associated set(1)/clear(0) were intended to
disable the RF output ports. However, set was used to indicate
that the port was being disabled.

That was opposite from the specsheet.

Renamed disableA/B to enableA/B and changed the logic so that
set was used to enable the port and clear disables the port.