Refactor default and working register arrays
Originally there was a haphazzard application of copying Curr.Reg
to defaultCurr and in other cases from defaultCurr to Curr.Reg.
This was difficult to track and in some cases required making a
third copy and wasting memory.

The solution I chose was to make a defaultRegisters array that was
read-only and all other actions would be performed on the working
registers, Curr.Reg[].

TESTING:
Allows for monitoring the changes any function
might make to the working registers while still having something
to compare the changes to.

RELEASE:
Allows for the creation of a reset function for use by end users.