Add comments describing test vs pio run
The use of negative logic to include/exclude unity_config.cpp was
confusing. When the file is included, which only happens when tests
are not being run, the line that includes unity_config.h makes it
seem as though tests are being re-included.

The extra comments clarify that unity_config.h is only there to
implement any per-project macros that you may have defined. If you
don't have any then that line is misleading/redundant.

You can just ignore it.