Debug No RF output after verifying exact data for registers
After verifying that the data, clock, and latch signals were being
sent to the max2871 the output was still not showing up on the
o'scope.

Reviewing the rest of the control signals lead to the realization
that the RF Output pin was not being set high which was failing to
enable the RF Output.