Refactor updateRegisters to perform clean-clock startup
The spec sheet defines a programming sequence that ensures what it calls a
clean-clock startup. (See pg. 13 of spec sheet)

According to the spec sheet:
Upon power-up, the registers should be programmed twice with at least
a 20ms pause between writes. The first write ensures that the device
is enabled, and the second write starts the VCO selection process.