Refactor updateRegisters to include clean-clock startup
When the max2871 IC is initialized the manufacturer provides a method for
getting a clean startup on the internal clock. This is skipped for all
subsequent chip programming updates.

Also, the DIVA bits in register 4 are checked for changes which will force
register 0 to buffer the write for register 4.

We don't check the bits in register 1 but if it changes at all its write
gets buffered by register 0.

By buffering the writes it prevents spurious changes on the RF outputs. For
example, changing the DIVA, vco divider, without buffering will cause the
RF out to change any time DIVA gets updated.