Add operational hardware tests
The project files contain 2 PlatformIO environments for the Spectrum
Analyzer control and RF boards.

Under env:uno we test the hardware with a combination of PlatformIO's
Unity test framework and Arduino calls.

Under env:smoke we test as though we are using the Arduino IDE. The sketch,
ci_smoke.ino can make standard library calls to the max2871 driver through
the ArduinoHAL.