Search actual hardware operation for Reg0 output only
There are situations where registers 4, 1, and 0 need to be programmed,
others where registers 1 and 0 or 4 and 0 need programming, and there are
just a few where only register 0 will be programmed.

This loop() was for finding those rare frequencies where only reg0 needed
programming so the proper operation of the code could be verified.